fix: logic bugs across frontend and backend
u16 edge index overflow for large graphs (>16k edge quads), get_commits/stream_commits
bypassing the AppState repo cache, misnamed search_file_path that searched commit
messages instead of file paths, stale CommitList selectedIdx after filtering, Escape
key double-firing between SearchBar and page handler, dead nullish coalescing on
missing translation keys, unhandled clipboard promise rejections, and missing
try/catch in browseForRepo dialogs.

- Change edge indices from u16 to u32 in render.rs and wgpu renderer
- Add Clone for GixRepository, use state.get_repo() in commits commands
- Rename search_file_path to search_message_contains with doc comment
- Add selectedIdx = 0 fallback when commit filtered out
- Add e.stopPropagation() in SearchBar Escape handler
- Fix translation fallback with proper key-exists check
- Add copyToClipboard helper with .catch() + toast in both components
- Wrap browseForRepo/browseForRepoNewWindow in try/catch, await openInNewWindow

+++ src-tauri/src/commands/render.rs
@@ -243,10 +243,10 @@ fn render_inner(wgpu_state: &WgpuState, input: &RenderGraphInput) -> Result<Vec<
     // For each 4-vertex quad generate 6 indices (2 triangles) for TriangleList.
     // Quad vertices: [left@start, right@start, left@end, right@end].
     // Triangles: [0,1,2] + [2,1,3] (both CCW under Y-down projection).
-    let edge_indices: Vec<u16> = (0..edge_vertices.len())
+    let edge_indices: Vec<u32> = (0..edge_vertices.len())
         .step_by(4)
         .flat_map(|base| {
-            let i = base as u16;
+            let i = base as u32;
             [i, i + 1, i + 2, i + 2, i + 1, i + 3]
         })
         .collect();
